requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
natsort>=8.4.0
alibabacloud_alimt20181012==1.1.0
